test: move webhook regressions out of generated tests
